+++ concrete_datastore/api/v1/permissions.py
@@ -1,6 +1,8 @@
 # coding: utf-8
 from importlib import import_module
-
+import logging
+import warnings
+from copy import deepcopy
 from django.db.models import Q
 from django.conf import settings
 from django.contrib.auth import get_user_model
@@ -8,12 +10,16 @@ from django.contrib.auth import get_user_model
 from rest_framework.exceptions import APIException
 from rest_framework import permissions, status
 
+import concrete_datastore.concrete.models
+from concrete_datastore.concrete.meta import meta_models
 from concrete_datastore.concrete.models import (
     ConcretePermission,
+    InstancePermission,
     DIVIDER_MODEL,
     UNDIVIDED_MODEL,
 )
 
+logger = logging.getLogger(__name__)
 
 DIVIDER_MODELs = "{}s".format(DIVIDER_MODEL)
 DIVIDER_MODELs_LOWER = DIVIDER_MODELs.lower()
@@ -253,20 +259,19 @@ class UserAccessPermission(permissions.BasePermission):
             else:
                 return False
         else:
-            return authenticated and (
-                (obj.created_by is not None and obj.created_by.pk == user.pk)
-                or (user.is_at_least_admin)
-                or (
-                    obj.can_admin_users.filter(pk=user.pk).exists()
-                    or does_intersect(
-                        obj.can_admin_groups, user.concrete_groups
-                    )
+            if authenticated is False:
+                return False
+            if user.is_at_least_admin:
+                return True
+
+            try:
+                model_name = obj._meta.model.__name__
+                instance_permissions = InstancePermission.objects.get(
+                    user_id=user.pk, model_name=model_name
                 )
-                or (
-                    at_least_staff
-                    and self.check_divider_permission(request, obj) is True
-                )
-            )
+            except InstancePermission.DoesNotExist:
+                return False
+            return str(obj.pk) in instance_permissions.write_instance_uids
 
 
 def get_available_scope_pks_for(user):
@@ -288,6 +293,10 @@ def apply_scope_filters(user, queryset):
 
 
 def filter_queryset_by_divider(queryset, user, divider):
+    warnings.warn(
+        'The method "filter_queryset_by_permissions" is deprecated',
+        DeprecationWarning,
+    )
     """
     queryset is the one of the current view,
     user is the connected one or anonymous
@@ -312,6 +321,10 @@ def filter_queryset_by_divider(queryset, user, divider):
 
 
 def filter_queryset_by_permissions(queryset, user, divider):
+    warnings.warn(
+        'The method "filter_queryset_by_permissions" is deprecated',
+        DeprecationWarning,
+    )
     if queryset.model == get_user_model():
         raise ValueError(
             "Queryset of model User cannot be filtered by permissions"
@@ -374,3 +387,343 @@ def filter_queryset_by_permissions(queryset, user, divider):
         queryset = queryset.distinct()
 
     return queryset
+
+
+def filter_queryset_by_permissions_and_scope(queryset, user, divider=None):
+    if user.is_anonymous:
+        return queryset.filter(public=True)
+    model_name = queryset.model.__name__
+    kwargs = {}
+    if divider is not None:
+        kwargs[DIVIDER_MODEL_LOWER] = divider
+    if user.is_authenticated and user.is_at_least_admin:
+        qs = queryset
+    else:
+        try:
+            instance_permission = user.instance_permissions.get(
+                model_name=model_name
+            )
+            filters = Q(pk__in=instance_permission.read_instance_uids) | Q(
+                public=True
+            )
+        except InstancePermission.DoesNotExist:
+            filters = Q(public=True)
+        qs = queryset.filter(filters)
+    if user.is_at_least_staff:
+        model_is_divided = model_name not in UNDIVIDED_MODEL
+        if model_is_divided:
+            divider_query = "{}_id__in".format(DIVIDER_MODEL_LOWER)
+            all_divider_pk = get_available_scope_pks_for(user)
+            qs |= queryset.filter(**{divider_query: all_divider_pk})
+
+    return qs.filter(**kwargs)
+
+
+def get_read_write_permission_users(
+    user,
+    instances_qs,
+    user_groups_pks,
+    include_divider=True,
+    include_view_groups=True,
+    include_admin_groups=True,
+    include_view_users=True,
+    include_admin_users=True,
+    user_level=None,
+):
+    if user_level is None:
+        user_level = user.level
+    user_id = user.pk
+    created_by_qs = instances_qs.filter(created_by_id=user_id).values_list(
+        'pk', flat=True
+    )
+    can_view_users_qs = instances_qs.filter(
+        can_view_users__pk=user_id
+    ).values_list('pk', flat=True)
+    can_view_groups_qs = instances_qs.filter(
+        can_view_groups__pk__in=user_groups_pks
+    ).values_list('pk', flat=True)
+
+    can_admin_users_qs = instances_qs.filter(
+        can_admin_users__pk=user_id
+    ).values_list('pk', flat=True)
+    can_admin_groups_qs = instances_qs.filter(
+        can_admin_groups__pk__in=user_groups_pks
+    ).values_list('pk', flat=True)
+    write_instances_uids = created_by_qs
+
+    if include_admin_groups is True:
+        write_instances_uids = write_instances_uids.union(can_admin_groups_qs)
+
+    if include_admin_users is True:
+        write_instances_uids = write_instances_uids.union(can_admin_users_qs)
+
+    #: A manager has write permissions on his scope
+    model_name = instances_qs.model.__name__
+    if (
+        (user_level == 'manager')
+        and (model_name not in UNDIVIDED_MODEL)
+        and (include_divider is True)
+    ):
+        all_divider_pk = getattr(user, DIVIDER_MODELs_LOWER).values_list(
+            'pk', flat=True
+        )
+        divider_field_pk = f"{DIVIDER_MODEL_LOWER}_id__in"
+        divided_qs = instances_qs.filter(
+            **{divider_field_pk: all_divider_pk}
+        ).values_list('pk', flat=True)
+        write_instances_uids = write_instances_uids.union(divided_qs)
+
+    read_instances_uids = write_instances_uids
+    if include_view_groups is True:
+        read_instances_uids = read_instances_uids.union(can_view_groups_qs)
+    if include_view_users is True:
+        read_instances_uids = read_instances_uids.union(can_view_users_qs)
+
+    write_instances_uids_set = set(
+        map(str, set(filter(lambda x: x is not None, write_instances_uids)))
+    )
+    read_instances_uids_set = set(
+        map(str, set(filter(lambda x: x is not None, read_instances_uids)))
+    )
+    return read_instances_uids_set, write_instances_uids_set
+
+
+def add_or_remove_element_into_list(
+    elt, current_list, check_list, has_changes=False
+):
+    """
+    Adds or removes an element from a list:
+      - if the element exists in the check list, and is not in the current_list
+        it is added to the curent_list
+      - if the element does not exist in the check_list and exists in the
+        current_list, it should be removed
+
+    example 1:
+    - elt : 5
+    - current_list : [1, 2, 3, 4, 5]
+    - check_list : [1, 2, 3]
+    => Result: the element 5 should be removed from the current_list
+
+    example 2:
+    - elt : 5
+    - current_list : [1, 2, 3, 4]
+    - check_list : [1, 2, 3, 5]
+    => Result: the element 5 should be added to the current_list
+
+    Arguments
+    ---------
+    :elt: the element to add/remove
+    :current_list: the list that contains the initial elements, and will
+        contain the final result
+    :check_list:: the list used to check whether the element should be added
+        or removed from the list
+    :has_changes: a boolean that is called recursively by the main method
+        to describe whether the data changed
+
+    Returns
+    -------
+    :has_changes: a boolean that describes whether the data has changed
+    :current_list: the list of the updated elements
+    """
+    if elt in current_list and elt not in check_list:
+        has_changes = True
+        current_list.remove(elt)
+    elif elt not in current_list and elt in check_list:
+        has_changes = True
+        current_list.append(elt)
+    return has_changes, current_list
+
+
+def update_instance_permission_uids(
+    perm_instance, new_read_instance_uids, new_write_instance_uids, all_uids
+):
+    """
+    Given a list of all_uids, this methods checks if the permission instance
+    should be updated or not.
+    If an element exists in both the all_uids and in the
+    new_read/write_instance_uids, it should also exist in the read/write uids
+    of the permission instance.
+    On the other side, if an element exists in the all_uids but does not exist
+    in the new_read/write_instance_uids, it should not exist in the read/write
+    uids of the permission instance.
+    The read_/write uids of the permission instance that do not appear in
+    all_uids should not be removed
+
+    Example:
+    if all_uids is [1, 2, 3, 4]
+    the read uids of the instance are [1, 2, 5, 6]
+    and the new read uids are [1, 3, 4]
+    We can see that the element 2 exists in all_uids and the read uids of the
+    permission instance, but not in the new read uids, so it should be removed
+    The elements 5 and 6 are in the read uid of the permission instance, but
+    are neither in the new read uids, nor in all_uids, so they should not be
+    removed. The final result of the red_uids are [1, 3, 4, 5, 6]
+    """
+    read_instance_uids = deepcopy(perm_instance.read_instance_uids)
+    write_instance_uids = deepcopy(perm_instance.write_instance_uids)
+    read_permission_changed = False
+    write_permission_changed = False
+    for uid in all_uids:
+        uid_str = str(uid)
+        #: Handle read instance uids
+        (
+            read_permission_changed,
+            read_instance_uids,
+        ) = add_or_remove_element_into_list(
+            elt=uid_str,
+            current_list=read_instance_uids,
+            check_list=new_read_instance_uids,
+            has_changes=read_permission_changed,
+        )
+
+        #: Handle write instance uids
+        (
+            write_permission_changed,
+            write_instance_uids,
+        ) = add_or_remove_element_into_list(
+            elt=uid_str,
+            current_list=write_instance_uids,
+            check_list=new_write_instance_uids,
+            has_changes=write_permission_changed,
+        )
+
+    if read_permission_changed or write_permission_changed:
+        if read_permission_changed:
+            perm_instance.read_instance_uids = read_instance_uids
+        if write_permission_changed:
+            perm_instance.write_instance_uids = write_instance_uids
+
+        logger.debug(
+            f'Updating permissions for {perm_instance.user.email} on '
+            f'instance <{all_uids.model.__name__}>'
+        )
+        return perm_instance
+        # perm_instance.save()
+
+
+def create_or_update_instance_permission_per_user(
+    user,
+    instances_qs,
+    include_divider=True,
+    include_view_groups=True,
+    include_admin_groups=True,
+    include_view_users=True,
+    include_admin_users=True,
+    user_level=None,
+):
+    if instances_qs.exists() is False:
+        return None, False
+    if user_level is None:
+        user_level = user.level
+    if user_level in ('superuser', 'admin'):
+        return None, False
+    model_name = instances_qs.model.__name__
+    user_groups_pks = user.concrete_groups.values_list('pk', flat=True)
+    (
+        read_instances_uids,
+        write_instances_uids,
+    ) = get_read_write_permission_users(
+        user=user,
+        instances_qs=instances_qs,
+        user_groups_pks=user_groups_pks,
+        include_divider=include_divider,
+        include_view_groups=include_view_groups,
+        include_admin_groups=include_admin_groups,
+        include_view_users=include_view_users,
+        include_admin_users=include_admin_users,
+        user_level=user_level,
+    )
+    should_create = False
+    try:
+        perm_instance = InstancePermission.objects.get(
+            user=user, model_name=model_name
+        )
+    except InstancePermission.DoesNotExist:
+        perm_instance = InstancePermission(user=user, model_name=model_name)
+        should_create = True
+
+    perm_instance = update_instance_permission_uids(
+        perm_instance=perm_instance,
+        new_read_instance_uids=read_instances_uids,
+        new_write_instance_uids=write_instances_uids,
+        all_uids=instances_qs.values_list('pk', flat=True),
+    )
+    return perm_instance, should_create
+
+
+def update_created_by_permissions(instance, user):
+    #: The creator of the instance has the read and write access
+    #: if the minimal levels on the datamodel allow it
+    model = instance._meta.model
+    model_name = model.__name__
+    has_read_permission = check_minimum_level('GET', user, model)
+    has_write_permission = check_minimum_level('PATCH', user, model)
+    defaults = {}
+    if has_read_permission is True:
+        defaults['read_instance_uids'] = [str(instance.pk)]
+    if has_write_permission is True:
+        defaults['write_instance_uids'] = [str(instance.pk)]
+    if not defaults:
+        return
+    permission_instance, created = InstancePermission.objects.get_or_create(
+        user=user, model_name=model_name, defaults=defaults
+    )
+    if created is False:
+        for field_name, field_value in defaults.items():
+            getattr(permission_instance, field_name).extend(field_value)
+        permission_instance.save()
+
+
+def check_instance_permissions_per_user(user, user_level=None):
+    if user_level is None:
+        user_level = user.level
+    if user_level in ('superuser', 'admin'):
+        return [], []
+    #: Checks the user permissions for all instances of the platform
+    instances_to_create = []
+    instances_to_update = []
+    for meta_model in meta_models:
+        model_name = meta_model.get_model_name()
+        if model_name in ('User', 'Group', 'Email'):
+            continue
+        concrete_model = getattr(
+            concrete_datastore.concrete.models, model_name
+        )
+        queryset = concrete_model.objects.all()
+        logger.debug(
+            f'Checking permission for user {user} on {queryset.count()} '
+            f'instances of model {model_name}'
+        )
+        (
+            instance,
+            should_create,
+        ) = create_or_update_instance_permission_per_user(
+            user=user, instances_qs=queryset, user_level=user_level
+        )
+        if instance is None:
+            continue
+        if should_create is True:
+            instances_to_create.append(instance)
+        else:
+            instances_to_update.append(instance)
+    return instances_to_create, instances_to_update
+
+
+def bulk_create_permission_instances(instances):
+    if not instances:
+        return
+    logger.info(f'Creating {len(instances)} Permission objects')
+    InstancePermission.objects.bulk_create(
+        objs=instances, batch_size=settings.BATCH_SIZE_FOR_BULK_OPERATIONS
+    )
+
+
+def bulk_update_permission_instances(instances):
+    if not instances:
+        return
+    logger.debug(f'Updating {len(instances)} Permission objects')
+    InstancePermission.objects.bulk_update(
+        objs=instances,
+        fields=['read_instance_uids', 'write_instance_uids'],
+        batch_size=settings.BATCH_SIZE_FOR_BULK_OPERATIONS,
+    )